requirements.txt
sec-api>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.0
psycopg2-binary>=2.9.0
yfinance>=0.2.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
openai>=1.0.0
polygon-api-client>=1.0.0
gunicorn>=21.2.0
voyageai>=0.3.0
pinecone-client>=3.0.0
tiktoken>=0.5.0
werkzeug>=3.0.0
